refactor: REFACTOR like seriously
