+++ lib/game_stats.rb
@@ -1,45 +1,11 @@
-require 'csv'
+require_relative 'stat_tracker'
 
-class GameData 
-  attr_reader :home_wins,
-              :away_wins,
-              :ties,
-              :total_games,
-              :total_goals,
-              :home_scores,
-              :away_scores
-
-  attr_accessor :games
-
+class GameStats < StatTracker
+  
   def initialize
-    @games = []
-    @home_wins = 0
-    @away_wins = 0
-    @ties = 0
-    @total_games = 0
-    @total_goals = 0
-    @home_scores = Hash.new(0)
-    @away_scores = Hash.new(0)
-    @teams = []
-    @lowest_scoring_away_team = nil
-    @lowest_scoring_home_team = nil
-  end
-
-  def add_games 
-    games = CSV.open './data/games.csv', headers: true, header_converters: :symbol
-    games.each do |row|
-      id = row[:game_id]
-      season = row[:season]
-      type = row[:type]
-      date_time = row[:date_time]
-      away = row[:away_team_id]
-      home = row[:home_team_id]
-      away_goals = row[:away_goals]
-      home_goals = row[:home_goals]
-      venue = row[:venue]
-      venue_link = row[:venue_link]
-      @games << Game.new(id, season, type, date_time, away, home, away_goals, home_goals, venue, venue_link)
-    end
+    @games = super
+    @teams = super
+    @game_teams = super
   end
 
   def wins_losses
@@ -60,7 +26,7 @@ class GameData
   def percentage_home_wins
     (@home_wins.to_f / @total_games).round(2)
   end
- 
+
   def percentage_visitor_wins
     (@away_wins.to_f / @total_games).round(2)
   end
@@ -78,7 +44,7 @@ class GameData
   def highest_total_score
     total_scores.last
   end
-  
+
   def lowest_total_score
     total_scores.first
   end
@@ -138,7 +104,7 @@ class GameData
       scores.max_by{|k,v| v}[0]
     end
   end
-  
+
   def highest_away_avg_score
     scores = Hash.new(0)
     @games.each do |game|
@@ -146,44 +112,11 @@ class GameData
       scores.max_by{|k,v| v}[0]
     end
   end
-  
+
   def team_goals
     @games.each do |game|
       @away_scores[game.away] += game.away_goals.to_i
       @home_scores[game.home] += game.home_goals.to_i
-    end
-  end
-
-  def lowest_scoring_away
-    lowest_scoring_team = @away_scores.min_by{|k, v| v}
-    @teams.each do |team|
-      if team.team_id == lowest_scoring_team[0]
-        @lowest_scoring_away_team = team.teamname
-      end
-    end
-    @lowest_scoring_away_team
-  end
-
-  def lowest_scoring_home
-    lowest_scoring_team = @home_scores.min_by{|k, v| v}
-    @teams.each do |team|
-      if team.team_id == lowest_scoring_team[0]
-        @lowest_scoring_home_team = team.teamname
-      end
-    end
-    @lowest_scoring_home_team
-  end
-
-  def add_teams
-    teams = CSV.open './data/teams.csv', headers: true, header_converters: :symbol
-    teams.each do |row|
-      team_id = row[:team_id]
-      franchiseid = row[:franchiseid]
-      teamname = row[:teamname]
-      abbreviation = row[:abbreviation]
-      stadium = row[:stadium]
-      link = row[:link]
-      @teams << Team.new(team_id, franchiseid, teamname, abbreviation, stadium, link)
     end
   end
 end
