+++ lib/game_team.rb
@@ -1,5 +1,3 @@
-require 'csv'
-
 class GameTeam
   attr_reader :game_id,
               :team_id,
